Ninja: Configure and build an helloworld program.
With this patch cmake with the Ninja generator is able to configure and
build a simple helloworld program.

Many targets are missing: all, install, clean, ... Basically there
is only the executable target.

The -rdynamic option is missing too.

Most of the flags are not handled yet.

All dependencies are not covered yet. Specially dependencies that triggers
re-configuration.