Partially fix link rules
Previously, we would attempt to add all link libraries to the explicit
dep list.  This is wrong, because this includes -l libraries,
which don't belong in the dep list.  If we added only non- -l deps
to the explicit dep list, we'd mess up the link order, which may
be significant for some builds.  Furthermore, not all targets put
$LINK_LIBRARIES right after $OBJECTS, possibly for good reasons.
Instead, add the library dependency list as implicit deps and pass
a full $LINK_LIBRARIES to the link rule.

This allows the CHECK_SYMBOL_EXISTS and various other tests performed
by cmake's own build system to pass, causing the configuration to
succeed.  There's still a remaining issue of absolute vs relative
paths which prevents the resultant build.ninja from working.