Re-work target ordering
Now we use order-only dependencies to enforce target ordering.
We also add all target dependencies, not just utilities (the only
reason we really need them is that in some build systems they are
used to mask utility dependencies).  We also resolve the dependency
to the actual files to avoid depending on a phony target, which is
suboptimal, since phony targets are always out of date.  .