Implement assumed source dependencies
If a source file is GENERATED and no custom command specifies it as
an output, assume that one of the target dependencies, OBJECT_DEPENDS
or header file custom commands will rebuild the file, and write a
phony rule causing them to be built first.  TODO: mark the phony rule
as restat.