Introduce new scheme for custom command target dependencies
Now, we look at all targets to which the custom command and take the
intersection of their target dependencies.

This should work, if we assume all target dependencies are a superset
of the true target dependencies for this custom command.  One can
imagine scenarios where they are not, mainly involving indirect
dependencies.  However this seems unlikely to arise in practice.

At the same time, move a bunch of code to cmLocalNinjaGenerator.